# crates/crucible-web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Crucible web API",
    "description": "The HTTP API that `cru web` serves. This document is generated from the axum router; do not edit it by hand.",
    "version": "0.30.0"
  },
  "paths": {
    "/api/agents": {
      "get": {
        "summary": "ACP agent profiles with probed availability, for the session-creation\nagent picker.",
        "description": "Served through the SWR catalog cache — the daemon probe takes ~0.5s and\nmust not gate every splash render.",
        "operationId": "list_agents",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentListResponse"
                }
              }
            }
          },
          "502": {
            "description": "The daemon could not list the agent profiles"
          }
        }
      }
    },
...
    "/api/providers": {
      "get": {
        "summary": "Served through the SWR catalog cache — provider probing takes ~0.7s and\nmust not gate every splash render. Shape: `{providers: [ProviderInfo]}`.",
        "description": "Takes no `kiln` parameter. It used to accept `kiln: Option<PathBuf>` and\nforward the raw directory to the daemon, which fed it to\n`find_workspace_and_resolve_classification` — so an arbitrary directory\ncould influence which providers a caller was told about, an input door\nstanding outside the registry floor every other kiln input now passes\nthrough. Nothing ever sent it (`listProviders()` takes no argument), so\nconverting it to a name would have preserved a door for no caller.",
        "operationId": "list_providers",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProvidersResponse"
                }
              }
            }
          },
          "502": {
            "description": "The daemon could not list the providers"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentListResponse": {
        "type": "object",
        "description": "The agent picker's list.\n\nThe key is `agents`, where the daemon's own answer says `profiles`: this\nroute names the list after what the picker shows, and the ROW is the\ndaemon's, so a probed field cannot be dropped on the way through.",
        "required": [
          "agents"
        ],
        "properties": {
          "agents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentProfileEntry"
            }
          }
        }
      },
...
      "AgentProfileEntry": {
        "type": "object",
        "description": "One ACP agent profile, with the availability probe's verdict.",
        "required": [
          "name",
          "description",
          "command",
          "is_builtin",
          "available"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "description": "The profile's description, or an empty string when it declares none."
          },
          "command": {
            "type": "string",
            "description": "The command that spawns the agent, or an empty string when the profile\nnames none. A profile with no command can never spawn, so it is never\navailable."
          },
          "is_builtin": {
            "type": "boolean",
            "description": "Whether the daemon ships this profile, rather than a config declaring it."
          },
          "available": {
            "type": "boolean",
            "description": "Whether the probe found the command on PATH and it answered `--version`."
          }
        }
      },
...
      "ProviderRow": {
        "type": "object",
        "description": "One LLM provider the daemon found.\n\nMirrors `crucible_core::types::ProviderInfo` field for field. It is\ndeclared here rather than re-exported because `crucible-core` takes no\nutoipa dependency, and a schema is what puts the fields in the document.",
        "required": [
          "name",
          "provider_type",
          "available",
          "models",
          "is_local"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "provider_type": {
            "type": "string",
            "description": "The backend behind the provider, such as `ollama` or `openai`."
          },
          "available": {
            "type": "boolean",
            "description": "Whether the provider answered its probe."
          },
          "default_model": {
            "type": [
              "string",
              "null"
            ]
          },
          "models": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "endpoint": {
            "type": [
              "string",
              "null"
            ]
          },
          "reason": {
            "type": [
              "string",
              "null"
            ],
            "description": "Why the provider is unavailable, when it is."
          },
          "is_local": {
            "type": "boolean",
            "description": "Whether the provider runs on this machine."
          }
        }
      },
      "ProvidersResponse": {
        "type": "object",
        "description": "What `GET /api/providers` answers.",
        "required": [
          "providers"
        ],
        "properties": {
          "providers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProviderRow"
            }
          }
        }
      },
...
    }
  }
}